requirements.txt
asgiref==3.2.7
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
CodeConvert==3.0.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.4
cryptography==2.8
Django==3.0.5
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-models-ext==1.1.9
django-rest-swagger==2.2.0
django-short-url==1.1.4
django-six==1.0.4
djangorestframework==3.11.0
Faker==0.9.1
furl==2.1.0
idna==2.9
importlib-metadata==1.6.0
itypes==1.1.0
Jinja2==2.11.1
jwt==1.0.0
MarkupSafe==1.1.1
mixer==6.1.3
mock==4.0.2
more-itertools==8.2.0
openapi-codec==1.3.2
orderedmultidict==1.0.1
packaging==20.3
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
screen==1.0.1
shortuuid==1.0.1
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.2
TimeConvert==1.5.1
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
xlwt==1.3.0
zipp==3.1.0
